Auto-git/gitscheduler/requirements.txt
click==8.1.7
GitPython==3.1.40
APScheduler==3.10.4
python-dateutil==2.8.2
colorama==0.4.6